+++ source/dotnet/Measurements.UnitTests/Application/Responses/GetAggregatedMeasurementsResponseTests.cs
@@ -19,23 +19,25 @@ public class GetAggregatedMeasurementResponseTests
 
         var qualities = new[] { "measured" };
         var resolutions = new[] { "PT1H" };
+        var units = new[] { "kWh" };
 
         var expectedDate = minObservationTime.ToDateOnly();
 
-        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>()
+        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>
         {
-            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions)),
+            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions, units)),
         };
 
         // Act
-        var result = GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements);
+        var actual = GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements);
 
         // Assert
-        var firstAggregation = result.MeasurementAggregations.First();
+        var firstAggregation = actual.MeasurementAggregations.First();
         Assert.Equal(expectedDate, firstAggregation.Date);
         Assert.Equal(100.0m, firstAggregation.Quantity);
         Assert.Equal(Quality.Measured, firstAggregation.Quality);
         Assert.False(firstAggregation.MissingValues);
+        Assert.False(firstAggregation.ContainsUpdatedValues);
     }
 
     [Fact]
@@ -46,20 +48,106 @@ public class GetAggregatedMeasurementResponseTests
         var maxObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow.AddHours(23));
         var qualities = new[] { "measured", "measured" };
         var resolutions = new[] { "PT1H", "PT15M" };
+        var units = new[] { "kWh" };
 
-        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>()
+        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>
         {
-            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions)),
+            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions, units)),
         };
 
         // Act & Assert
         Assert.Throws<InvalidOperationException>(() => GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements));
     }
 
-    /* TODO: Test for missing values, invalid qualities, and other edge cases */
+    [Fact]
+    public void Create_MultipleUnits_ShouldThrowException()
+    {
+        // Arrange
+        var minObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow);
+        var maxObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow.AddHours(23));
+        var qualities = new[] { "measured", "measured" };
+        var resolutions = new[] { "PT1H" };
+        var units = new[] { "kWh", "kVArh" };
+
+        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>
+        {
+            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions, units)),
+        };
+
+        // Act & Assert
+        Assert.Throws<InvalidOperationException>(() => GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements));
+    }
+
+    [Fact]
+    public void Create_DataContainsMissingValues_ShouldSetMissingValuesToTrue()
+    {
+        // Arrange
+        var minObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow);
+        var maxObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow.AddHours(24)); // 24 hours to simulate missing values
+        var qualities = new[] { "measured" };
+        var resolutions = new[] { "PT1H" };
+        var units = new[] { "kWh" };
+
+        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>
+        {
+            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions, units)),
+        };
+
+        // Act
+        var actual = GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements);
+
+        // Assert
+        Assert.True(actual.MeasurementAggregations.First().MissingValues);
+    }
+
+    [Fact]
+    public void Create_DataContainsInvalidQualities_ShouldThrowException()
+    {
+        // Arrange
+        var minObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow);
+        var maxObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow.AddHours(23));
+        var qualities = new[] { "invalid_quality" };
+        var resolutions = new[] { "PT1H" };
+        var units = new[] { "kWh" };
+
+        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>
+        {
+            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions, units)),
+        };
+
+        // Act & Assert
+        Assert.Throws<ArgumentOutOfRangeException>(() => GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements));
+    }
+
+    [Fact]
+    public void Create_DataContainsUpdatedObservations_ShouldSetContainsUpdatedValues()
+    {
+        // Arrange
+        var minObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow);
+        var maxObservationTime = Instant.FromDateTimeOffset(DateTimeOffset.UtcNow.AddHours(23));
+        var qualities = new[] { "measured" };
+        var resolutions = new[] { "PT1H" };
+        var units = new[] { "kWh" };
+
+        var aggregatedMeasurements = new List<AggregatedMeasurementsResult>
+        {
+            new(CreateRaw(minObservationTime, maxObservationTime, qualities, resolutions, units, 2L)),
+        };
+
+        // Act
+        var actual = GetAggregatedMeasurementsResponse.Create(aggregatedMeasurements);
+
+        // Assert
+        Assert.True(actual.MeasurementAggregations.First().ContainsUpdatedValues);
+    }
 
     private static ExpandoObject CreateRaw(
-        Instant minObservationTime, Instant maxObservationTime, string[] qualities, string[] resolutions)
+        Instant minObservationTime,
+        Instant maxObservationTime,
+        string[] qualities,
+        string[] resolutions,
+        string[] units,
+        long observationUpdates = 1L)
     {
         dynamic raw = new ExpandoObject();
         raw.min_observation_time = minObservationTime.ToDateTimeOffset();
@@ -67,7 +155,9 @@ public class GetAggregatedMeasurementResponseTests
         raw.aggregated_quantity = 100.0m;
         raw.qualities = qualities;
         raw.resolutions = resolutions;
+        raw.units = units;
         raw.point_count = 24L;
+        raw.observation_updates = observationUpdates;
 
         return raw;
     }