+++ source/dotnet/Measurements.UnitTests/Infrastructure/Persistence/Queries/GetAggregatedByMonthQueryTests.cs
@@ -8,20 +8,20 @@ using Xunit;
 
 namespace Energinet.DataHub.Measurements.UnitTests.Infrastructure.Persistence.Queries;
 
-public class GetAggregatedMeasurementsQueryTests
+public class GetAggregatedByMonthQueryTests
 {
     [Theory]
     [AutoData]
-    public void ToString_Returns_ExpectedResult(string meteringPointId, [Range(-9998, 9999)]int year, [Range(1, 12)]int month)
+    public void ToString_Returns_ExpectedResult(string meteringPointId, [Range(-9998, 9999)] int year, [Range(1, 12)] int month)
     {
         // Arrange
         var yearMonth = new YearMonth(year, month);
         var databricksSchemaOptions = new DatabricksSchemaOptions { CatalogName = "spark_catalog", SchemaName = "schema_name" };
         var expected = CreateExpectedQuery(databricksSchemaOptions);
-        var getAggregatedMeasurementsQuery = new GetAggregatedMeasurementsQuery(meteringPointId, yearMonth, databricksSchemaOptions);
+        var getAggregatedByMonthQuery = new GetAggregatedByMonthQuery(meteringPointId, yearMonth, databricksSchemaOptions);
 
         // Act
-        var actual = getAggregatedMeasurementsQuery.ToString();
+        var actual = getAggregatedByMonthQuery.ToString();
 
         // Assert
         Assert.Equal(expected, actual);
@@ -31,6 +31,7 @@ public class GetAggregatedMeasurementsQueryTests
     {
         return $"with most_recent as (" +
                $"select row_number() over (partition by {MeasurementsGoldConstants.MeteringPointIdColumnName}, {MeasurementsGoldConstants.ObservationTimeColumnName} order by {MeasurementsGoldConstants.TransactionCreationDatetimeColumnName} desc) as row, " +
+               $"count(*) over (partition by {MeasurementsGoldConstants.MeteringPointIdColumnName}, {MeasurementsGoldConstants.ObservationTimeColumnName}) as row_count, " +
                $"{MeasurementsGoldConstants.MeteringPointIdColumnName}, {MeasurementsGoldConstants.UnitColumnName}, {MeasurementsGoldConstants.ObservationTimeColumnName}, {MeasurementsGoldConstants.QuantityColumnName}, {MeasurementsGoldConstants.QualityColumnName}, {MeasurementsGoldConstants.ResolutionColumnName}, {MeasurementsGoldConstants.IsCancelledColumnName} " +
                $"from {databricksSchemaOptions.CatalogName}.{databricksSchemaOptions.SchemaName}.{MeasurementsGoldConstants.TableName} " +
                $"where {MeasurementsGoldConstants.MeteringPointIdColumnName} = :{QueryParameterConstants.MeteringPointIdParameter} " +
@@ -38,17 +39,19 @@ public class GetAggregatedMeasurementsQueryTests
                $"and {MeasurementsGoldConstants.ObservationTimeColumnName} < :{QueryParameterConstants.ObservationTimeToParameter} " +
                $") " +
                $"select {MeasurementsGoldConstants.MeteringPointIdColumnName}, " +
-               $"min({MeasurementsGoldConstants.ObservationTimeColumnName}) as {AggregatedMeasurementsConstants.MinObservationTime}, " +
-               $"max({MeasurementsGoldConstants.ObservationTimeColumnName}) as {AggregatedMeasurementsConstants.MaxObservationTime}, " +
-               $"sum({MeasurementsGoldConstants.QuantityColumnName}) as {AggregatedMeasurementsConstants.AggregatedQuantity}, " +
-               $"array_agg(distinct({MeasurementsGoldConstants.QualityColumnName})) as {AggregatedMeasurementsConstants.Qualities}, " +
-               $"array_agg(distinct({MeasurementsGoldConstants.ResolutionColumnName})) as {AggregatedMeasurementsConstants.Resolutions}, " +
-               $"count({MeasurementsGoldConstants.ObservationTimeColumnName}) as {AggregatedMeasurementsConstants.PointCount} " +
+               $"min({MeasurementsGoldConstants.ObservationTimeColumnName}) as {AggregatedQueryConstants.MinObservationTime}, " +
+               $"max({MeasurementsGoldConstants.ObservationTimeColumnName}) as {AggregatedQueryConstants.MaxObservationTime}, " +
+               $"sum({MeasurementsGoldConstants.QuantityColumnName}) as {AggregatedQueryConstants.AggregatedQuantity}, " +
+               $"array_agg(distinct({MeasurementsGoldConstants.QualityColumnName})) as {AggregatedQueryConstants.Qualities}, " +
+               $"array_agg(distinct({MeasurementsGoldConstants.ResolutionColumnName})) as {AggregatedQueryConstants.Resolutions}, " +
+               $"array_agg(distinct({MeasurementsGoldConstants.UnitColumnName})) as {AggregatedQueryConstants.Units}, " +
+               $"count({MeasurementsGoldConstants.ObservationTimeColumnName}) as {AggregatedQueryConstants.PointCount}, " +
+               $"max(row_count) as {AggregatedQueryConstants.ObservationUpdates} " +
                $"from most_recent " +
                $"where row = 1 " +
                $"and not {MeasurementsGoldConstants.IsCancelledColumnName} " +
                $"group by {CreateGroupByStatement()} " +
-               $"order by {AggregatedMeasurementsConstants.MinObservationTime}";
+               $"order by {AggregatedQueryConstants.MinObservationTime}";
     }
 
     private static string CreateGroupByStatement()